+++ src/containers/App.js
@@ -1,16 +1,15 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { showPaletteCodeModal, closePaletteCodeModal } from '../actions/modal';
-import { showSnackBarWithText, closeSnackBar } from '../actions/snackBar';
+import { showPaletteCodeModal } from '../actions/modal';
+import { closeSnackBar } from '../actions/snackBar';
 
 import App from '../components/App/App';
 
 
 function mapStateToProps(state) {
-  const { palette, modal, snackBar } = state;
+  const { palette, snackBar } = state;
   return {
     palette,
-    isVisibleCodeModal: modal.isVisible,
     isVisibleSnackBar: snackBar.isVisible,
     snackBarText: snackBar.text,
   };
@@ -20,8 +19,6 @@ function mapStateToProps(state) {
 function mapDispatchToProps(dispatch) {
   return bindActionCreators({
     showPaletteCodeModal,
-    closePaletteCodeModal,
-    showSnackBarWithText,
     closeSnackBar,
   }, dispatch);
 }
