+++ src/containers/CopyCodeModal.js
@@ -1,0 +1,29 @@
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import { closePaletteCodeModal } from '../actions/modal';
+import { showSnackBarWithText } from '../actions/snackBar';
+import { setPaletteColor } from '../actions/palette';
+import CopyCodeModal from '../components/CopyCodeModal/CopyCodeModal';
+import getAndroidTheme from '../lib/getAndroidTheme';
+
+function mapStateToProps(state) {
+  const { modal, palette } = state;
+  return {
+    contents: getAndroidTheme(palette),
+    isVisibleCodeModal: modal.isVisible,
+  };
+}
+
+
+function mapDispatchToProps(dispatch) {
+  return bindActionCreators({
+    setPaletteColor,
+    closePaletteCodeModal,
+    showSnackBarWithText,
+  }, dispatch);
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(CopyCodeModal);
